MLXPK-1821: Move model layers to core folder
