+++ scripts/safetensors_generator/dataset_reader.py
@@ -12,7 +12,7 @@ import numpy as np
 import torch
 
 from ng_model_gym.core.data.utils import tonemap_inverse
-from ng_model_gym.usecases.nss.model.graphics_utils import swizzle
+from ng_model_gym.core.model.graphics_utils import swizzle
 from scripts.safetensors_generator.exr_utils import (
     create_depth_params,
     read_exr_torch,
