+++ src/ng_model_gym/core/model/dense_warp_utils.py
