+++ src/ng_model_gym/core/model/graphics_utils.py
