+++ src/ng_model_gym/core/trainer/trainer.py
@@ -13,14 +13,14 @@ from tqdm.auto import tqdm
 
 from ng_model_gym.core.data.dataloader import get_dataloader
 from ng_model_gym.core.data.utils import DataLoaderMode
+from ng_model_gym.core.evaluator.metrics import get_metrics
 from ng_model_gym.core.loss.losses import LossV1
 from ng_model_gym.core.optimizers.lars_adam import lars_adam_torch
+from ng_model_gym.core.schedulers.lr_scheduler import CosineAnnealingWithWarmupLR
 from ng_model_gym.core.utils.checkpoint_utils import latest_checkpoint_path
 from ng_model_gym.core.utils.config_model import ConfigModel, TrainingConfig
 from ng_model_gym.core.utils.general_utils import create_directory
 from ng_model_gym.core.utils.types import LearningRateScheduler, TrainEvalMode
-from ng_model_gym.usecases.nss.metrics import get_metrics
-from ng_model_gym.usecases.nss.model.lr_scheduler import CosineAnnealingWithWarmupLR
 from ng_model_gym.usecases.nss.model.model import create_model
 from ng_model_gym.usecases.nss.model.model_v1 import QATNSSModel
 from ng_model_gym.usecases.nss.model.shaders.slang_utils import load_slang_module
