+++ tests/nss/unit/test_metrics.py
@@ -7,7 +7,7 @@ import torch
 from torchmetrics.functional.image import peak_signal_noise_ratio
 from torchmetrics.image.ssim import StructuralSimilarityIndexMeasure
 
-from ng_model_gym.usecases.nss.metrics import (
+from ng_model_gym.core.evaluator.metrics import (
     get_metrics,
     Psnr,
     RecPsnr,
